extract rpc-responses-tx-status
