+++ crates/rpc-responses-common/src/lib.rs
@@ -8,9 +8,8 @@ use pyo3::prelude::*;
 use serde::{Deserialize, Serialize};
 use serde_with::{serde_as, DisplayFromStr, TryFromInto};
 use solana_sdk::{
-    clock::{Epoch, Slot, UnixTimestamp},
+    clock::{Epoch, Slot},
     epoch_info::EpochInfo as EpochInfoOriginal,
-    transaction::TransactionError as TransactionErrorOriginal,
 };
 use solders_account::{Account, AccountJSON};
 use solders_account_decoder::{
@@ -20,12 +19,7 @@ use solders_account_decoder::{
 use solders_hash::Hash as SolderHash;
 use solders_macros::{common_methods, richcmp_eq_only, EnumIntoPy};
 use solders_pubkey::Pubkey;
-use solders_signature::Signature;
 use solders_transaction_error::TransactionErrorType;
-use solders_transaction_status::{
-    tmp_transaction_status::TransactionConfirmationStatus as TransactionConfirmationStatusOriginal,
-    TransactionConfirmationStatus,
-};
 
 use solders_rpc_common::response_data_boilerplate;
 
@@ -526,76 +520,6 @@ impl RpcKeyedAccountJsonParsed {
 pub enum RpcKeyedAccountMaybeJSON {
     Binary(RpcKeyedAccount),
     Parsed(RpcKeyedAccountJsonParsed),
-}
-
-// the one in solana_client uses transaction_status
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-struct RpcConfirmedTransactionStatusWithSignatureOriginal {
-    pub signature: String,
-    pub slot: Slot,
-    pub err: Option<TransactionErrorOriginal>,
-    pub memo: Option<String>,
-    pub block_time: Option<UnixTimestamp>,
-    pub confirmation_status: Option<TransactionConfirmationStatusOriginal>,
-}
-
-#[derive(Debug, PartialEq, Eq, Serialize, Deserialize, Clone, From, Into)]
-#[pyclass(module = "solders.rpc.responses", subclass)]
-pub struct RpcConfirmedTransactionStatusWithSignature(
-    RpcConfirmedTransactionStatusWithSignatureOriginal,
-);
-
-response_data_boilerplate!(RpcConfirmedTransactionStatusWithSignature);
-
-#[richcmp_eq_only]
-#[common_methods]
-#[pymethods]
-impl RpcConfirmedTransactionStatusWithSignature {
-    #[new]
-    pub fn new(
-        signature: Signature,
-        slot: Slot,
-        err: Option<TransactionErrorType>,
-        memo: Option<String>,
-        block_time: Option<UnixTimestamp>,
-        confirmation_status: Option<TransactionConfirmationStatus>,
-    ) -> Self {
-        RpcConfirmedTransactionStatusWithSignatureOriginal {
-            signature: signature.to_string(),
-            slot,
-            err: err.map(|e| e.into()),
-            memo,
-            block_time,
-            confirmation_status: confirmation_status.map(|c| c.into()),
-        }
-        .into()
-    }
-
-    #[getter]
-    pub fn signature(&self) -> Signature {
-        Signature::from_str(&self.0.signature).unwrap()
-    }
-    #[getter]
-    pub fn slot(&self) -> Slot {
-        self.0.slot
-    }
-    #[getter]
-    pub fn err(&self) -> Option<TransactionErrorType> {
-        self.0.err.clone().map(|e| e.into())
-    }
-    #[getter]
-    pub fn memo(&self) -> Option<String> {
-        self.0.memo.clone()
-    }
-    #[getter]
-    pub fn block_time(&self) -> Option<UnixTimestamp> {
-        self.0.block_time
-    }
-    #[getter]
-    pub fn confirmation_status(&self) -> Option<TransactionConfirmationStatus> {
-        self.0.confirmation_status.clone().map(|s| s.into())
-    }
 }
 
 // the one in solana_client uses account_decoder
