+++ crates/rpc-responses-tx-status/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "solders-rpc-responses-common"
+name = "solders-rpc-responses-tx-status"
 version = "0.14.4"
 edition = "2021"
 include = ["/src"]
@@ -9,17 +9,11 @@ repository = "https://github.com/kevinheavey/solders"
 [dependencies]
 pyo3 = { workspace = true }
 serde = { workspace = true }
-serde_with = { workspace = true }
 derive_more = { workspace = true }
 solders-macros = { workspace = true }
-solders-hash = { workspace = true }
 solders-traits-core = { workspace = true }
-solders-account = { workspace = true }
-solders-account-decoder = { workspace = true }
 solders-transaction-error = { workspace = true }
+solders-transaction-status = { workspace = true }
 solders-rpc-common = { workspace = true }
-solders-rpc-version = { workspace = true }
-solders-pubkey = { workspace = true }
 solders-signature = { workspace = true }
 solana-sdk = { workspace = true }
-camelpaste = { workspace = true }
