+++ crates/rpc-responses/Cargo.toml
@@ -22,6 +22,7 @@ solders-account-decoder = { workspace = true }
 solders-transaction-error = { workspace = true }
 solders-transaction-status = { workspace = true }
 solders-rpc-responses-common = { workspace = true }
+solders-rpc-responses-tx-status = { workspace = true }
 solders-rpc-common = { workspace = true }
 solders-rpc-errors = { workspace = true }
 solders-rpc-version = { workspace = true }
